refactor: rename vault hook and add additional api calls
